+++ src/app/app.module.ts
@@ -3,20 +3,25 @@ import { NgModule } from '@angular/core';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { FirstComponentComponent } from './first-component/first-component.component';
-import { ProductComponentComponent } from './product-component/product-component.component';
+import { FirstComponent } from './first-component/first-component.component';
+import { ProductComponent } from './product-component/product-component.component';
+import { DEFAULT_CURRENCY_CODE } from '@angular/core';
+import { ProductListComponent } from './product-list/product-list.component';
 
 @NgModule({
   declarations: [
     AppComponent,
-    FirstComponentComponent,
-    ProductComponentComponent
+    FirstComponent,
+    ProductComponent,
+    ProductListComponent,
   ],
   imports: [
     BrowserModule,
     AppRoutingModule
   ],
-  providers: [],
+  providers: [
+    {provide: DEFAULT_CURRENCY_CODE, useValue: 'EUR'},
+],
   bootstrap: [AppComponent]
 })
 export class AppModule { }