+++ src/app/first-component/first-component.component.ts
@@ -12,7 +12,7 @@ enum ProductCategory {
   templateUrl: './first-component.component.html',
   styleUrls: ['./first-component.component.scss']
 })
-export class FirstComponentComponent implements OnInit {
+export class FirstComponent implements OnInit {
   name: string;
   description: string;
   price: number;