+++ src/app/model/model.ts
@@ -1,0 +1,8 @@
+export interface ProductModel {
+    name: string;
+    amount: number;
+    priceMin: number;
+    priceMax: number;
+    currency?: string;
+    description?: string;
+}