+++ src/app/product-component/product-component.component.ts
@@ -1,11 +1,14 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
+import { ProductModel } from '../model/model';
 
 @Component({
   selector: 'app-product-component',
   templateUrl: './product-component.component.html',
   styleUrls: ['./product-component.component.scss']
 })
-export class ProductComponentComponent {
+export class ProductComponent {
+  @Input() product: ProductModel;
+
   onBuy() {
     console.log('Товар куплен');
   }